dynamic-layers: qt5: Fix build error with RZ/G1 when graphic feature available
When build core-image-qt for RZ/G1 with graphic feature available,
Yocto will raise error in do_package_qa because some packages
use libEGL or libGLESv2 but no provider in RDEPENDS.

The usual fix for this kind of error is adding RDEPENDS.
But it is not good in this case because there are 2 different provider
for these graphic libraries.
In case build with meta-renesas only, provider for graphic libraries is mesa
But if build with Renesas graphic feature package, provider is another recipe.

So use INSANE_SKIP to avoid the error is the correct fix.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>