rz-common: gst-plugins-bad: Depending bayer2raw if layer meta-rz-graphics is added
Currently, a build error occurs when building gstreamer-plugins-bad without the
meta-rz-features/meta-rz-graphics layer. This error is caused by the bayer2raw
recipe, which lacks the EGL/eglext_REL.h header file. To resolve this issue,
add a judgement to check whether the meta-rz-graphics layer is included or not.
If it is not included, exclude the bayer2raw, bayerconver and bayersink.

Note: This update apply for both G2 series and G2L series.

Signed-off-by: trungvanle <trung.le.xk@renesas.com>